docs: Convert slack links to CNCF, add link to contributing guide (#40)
**What changed?**

Converts links from our old slack instance to the CNCF community
inviter.
Adds a link to our shared contributing guide to the contributing readme.

**Why?**

The goal of the contribution guide is to move all common guides and
processes between our repositories to a shared place – reducing
conflicting information and streamlining the process to contribute to
cadence.

**How did you test it?**

N/A

Potential risks

N/A

Release notes

N/A

Documentation Changes

N/A