feat: Define workflow signal decorator and definition for registry (#46)
<!-- Describe what has changed in this PR -->
**What changed?**
Define workflow signal decorator and definition for registry

<!-- Tell your future self why have you made these changes -->
**Why?**
This is a must have component in order to implement signals in cadence
workflow. It allows user to define a method in their workflow class that
can respond to signal.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Unit tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: Tim Li <ltim@uber.com>